# w4/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FlyRank Auth API",
    "version": "1.0.0",
    "description": "Secure API with Supabase Auth — sign up, log in, log out, and protected routes."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local dev server" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "AuthBody": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email":    { "type": "string", "example": "test@example.com" },
          "password": { "type": "string", "example": "password123" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "summary": "Create a new user account",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthBody" }
            }
          }
        },
        "responses": {
          "201": { "description": "User created successfully" },
          "400": {
            "description": "Missing fields or signup error",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Authenticate and receive a JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthBody" }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful — returns access_token and refresh_token" },
          "400": { "description": "Missing fields" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "summary": "End the user's session",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "204": { "description": "Logged out successfully" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/public/info": {
      "get": {
        "tags": ["Public"],
        "summary": "Open public endpoint — no auth needed",
        "responses": {
          "200": { "description": "Public message" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "tags": ["Protected"],
        "summary": "Get the logged-in user's profile",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "User profile (id, email, created_at)" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "tags": ["Protected"],
        "summary": "Protected dashboard — same middleware, no new auth code",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard message" },
          "401": { "description": "Missing or invalid token" }
        }
      }
    }
  }
}